requirements.txt
chardet==4.0.0
chardet==5.2.0
fitz==0.0.1.dev2
jieba==0.42.1
Jinja2==3.1.6
numpy==2.3.1
pandas==2.3.0
pdfplumber==0.11.6
Pillow==11.3.0
python_docx==1.1.2
scikit_learn==1.7.0
tqdm==4.66.5
tqdm==4.67.1
unrar==0.4
